File Structure // Script Call Issue
As this project gets larger, organization becomes more important. As I attempted to reorganize the filing of different files, I ran into the problem of scripts like "keys.py" or "mouse.py"
not being able to be ran via "os.startfile()" if they were not in the same exact directory as the script calling on them. Unfortunately, this will lead to my gui files being in the same
files as all my other python scripts.